requirements.txt
ExifRead==2.1.2
numpy==1.12.1
opencv-python==3.2.0.7
python-magic==0.4.13